// osa9/flight-diary-frontend/src/types.ts
export enum Weather {
    Sunny = 'sunny',
    Rainy = 'rainy',
    Cloudy = 'cloudy',
    Stormy = 'stormy',
    Windy = 'windy',
  }

  export enum Visibility {
    Great = 'great',
    Good = 'good',
    Ok = 'ok',
    Poor = 'poor',
  }

  export interface DiaryEntry {
    id: number;
    date: string;
    weather: Weather;
    visibility: Visibility;
    comment: string;
  }

  export type NewDiaryEntry = Omit<DiaryEntry, 'id'>;

  export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>;